Refactor project to be a normal ROS package (#33)
* Move to more typical ROS package structure to allow for inclusion in the build process of other packages

* Do not expect a test to fail any more, since ROS was fixed by now

* Update CI to new versions and adapt it to the new layout

* Fix CI

* Debug CI

* More CI debugging

* Fix Sphinx configuration

* Revert debugging attempts

* Cleanup and attempt using CI tools

* debug

* Try without ros-tooling/setup-ros

* Fix installation of packages

* Install dev tools

* Run `rosdep init`

* Update dependency installation

* Fix dependency installation

* Switch from black to ruff

* Try to fix target-ros2-distro argument passing

* Parameterize test

* Ignore CI artifacts in formatting/linting

* Try to repair ruff excludes

* Actually install dependencies

* Fix some paths

* Debug again

* Cleanup

* Link to contributors

* Apply formatting

* Fix tests cases broken after merge

* Slightly better signal handling in with_launch_file

* Remove foxy leftovers

* Cleanup

+++ ros2_easy_test/ros2_easy_test/decorators.py
@@ -130,7 +130,8 @@ def with_single_node(
                     )
                     # The type ignore is needed due to a bug in ROS2 Humble+
                     executor.spin_until_future_complete(
-                        test_function_task, timeout_sec=time_limit  # type: ignore[arg-type]
+                        test_function_task,
+                        timeout_sec=time_limit,  # type: ignore[arg-type]
                     )
 
                     test_function_exception = test_function_task.exception()
@@ -294,7 +295,8 @@ def with_launch_file(  # noqa: C901
                         )
                         # The type ignore is needed due to a bug in ROS2 Humble+
                         executor.spin_until_future_complete(
-                            test_function_task, timeout_sec=time_limit  # type: ignore[arg-type]
+                            test_function_task,
+                            timeout_sec=time_limit,  # type: ignore[arg-type]
                         )
 
                         test_function_exception = test_function_task.exception()
@@ -341,16 +343,18 @@ def with_launch_file(  # noqa: C901
                     except NameError:
                         pass  # Maybe, the launch process was not even started
                     else:
-                        # Signal the child launch process to finish
-                        # This is much like pressing Ctrl+C on the console
-                        ros2_process.send_signal(SIGINT)
-                        try:
-                            # Might raise a TimeoutExpired if it takes too long
-                            return_code = ros2_process.wait(timeout=shutdown_timeout / 2)
-                        except TimeoutExpired:  # pragma: no cover
-                            ros2_process.terminate()
-                            # return_code will be larger than 130
-                            return_code = ros2_process.wait(timeout=shutdown_timeout / 2)
+                        return_code = ros2_process.poll()
+                        if return_code is None:
+                            # Signal the child launch process to finish
+                            # This is much like pressing Ctrl+C on the console
+                            ros2_process.send_signal(SIGINT)
+                            try:
+                                # Might raise a TimeoutExpired if it takes too long
+                                return_code = ros2_process.wait(timeout=shutdown_timeout / 2)
+                            except TimeoutExpired:  # pragma: no cover
+                                ros2_process.terminate()
+                                # return_code will be larger than 130
+                                return_code = ros2_process.wait(timeout=shutdown_timeout / 2)
 
                 # Both SUCCESS (0) or the result code of SIGINT (130) are acceptable
                 return_code_problematic = return_code not in {0, 130}
